+++ lib/main.dart
@@ -1,3 +1,5 @@
+import 'dart:io';
+
 import 'package:firebase_core/firebase_core.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
@@ -37,7 +39,10 @@ class MyApp extends ConsumerWidget {
           routes: [
             GoRoute(
               path: 'edit',
-              builder: (context, state) => const ImageEditingPage(),
+              builder: (context, state) {
+                final imageFile = state.extra as File;
+                return ImageEditingPage(imageFile: imageFile);
+              },
             ),
           ],
         ),